+++ src/components/dialog/content/ErrorDialogContent.vue
